Put the IOB fuzzer back in the tiled_fuzzer.
